Change import to xkcd as xkcd_api, fixes CSRF error recovery.
When reloading app, sometimes encounter a CSRF error that propagates as Flask can't find the /xkcd route (it was named xkcd_comic to avoid a name clash with the imported API module).
Fix this by renaming route function to xkcd, and importing APIs: api_name as api_name_api, where this is necessary.

This might not always be necessary iss_loc will probably be called iss or iss_locator in app, etc.