Add comic details to page title, refactor test.
Break out site title and page title into separate assertions.
Anything but the pure 'route title' and the site title element breaks
the test test_subpage_internal_routes. Breaking it up to test the site
title, and passing in the bare route (or whatever static text is passed
to the test) separately allows the title to be dynamic and still pass
the test.

NB xkcd title is changed to include comic title, and defaults to the
latest comic - so this will change over time.